ref(backup): Split out import/export functions
This will have two benefits:

1. It makes the code more testable, since "printing to the terminal" is
   now a thing tests can intercept and potentially assert against.
2. This allows us to mess with the signature of the underlying functions
   as we implement more features without exposing the changes to users
   and causing undue churn. As we add new filtering options, like
   `--scope` and `--slug`, we want to avoid "releasing" them piecemeal,
   and instead hide them behind the current interfaces.

Issue: getsentry/team-ospo#166
